requirements-cpu.txt
# Ant Tracker - CPU-only Installation
# Use this if you don't have an NVIDIA graphics card or want to use CPU processing
# Note: CPU processing is slower but works on all computers
#
# Installation Steps:
# 1. Install PyTorch CPU version first:
#    pip install torch torchvision --index-url https://download.pytorch.org/whl/cpu
#
# 2. Then install these packages:
#    pip install -r requirements-cpu.txt
#
# Expected processing speed:
#   - 10-minute video may take 30-60 minutes to process on CPU
#   - Training a model may take several hours

# Core computer vision and deep learning
ultralytics>=8.2.0         # YOLOv8 with oriented bounding box support
opencv-python>=4.8.0       # Image and video processing
numpy>=1.24.0,<2.0.0       # Numerical operations
pillow>=10.0.0             # Image handling

# Machine learning utilities
scikit-learn>=1.3.0        # Dataset splitting and preprocessing
pandas>=2.0.0              # Data analysis and CSV export
matplotlib>=3.7.0          # Plotting and visualization
seaborn>=0.12.0            # Statistical plots
tqdm>=4.65.0               # Progress bars

# Tracking algorithms
lap>=0.4.0                 # Linear Assignment Problem solver for BoT-SORT tracking

# Video processing
ffmpeg-python>=0.2.0       # Video encoding/decoding (requires ffmpeg installed)

# Configuration and utilities
PyYAML>=6.0                # YAML configuration files
psutil>=5.9.0              # System monitoring
tensorboard>=2.13.0        # Training visualization (optional)

# Note: PyTorch and torchvision should be installed separately using the command above
# This ensures you get the CPU-only version which is smaller and faster to download
